fixed after ccr

+++ src/main/java/oop/diskUsage/Measurement.java
@@ -3,7 +3,7 @@ package oop.diskUsage;
 import java.text.DecimalFormat;
 
 public class Measurement {
-    public static String printSizeOfFile(double size){
+    public static String printSizeOfFile(long size){
 
         DecimalFormat df = new DecimalFormat("#.###");
